backend/requirements.txt
fastapi>=0.115.5
uvicorn>=0.32.1
python-dotenv>=1.0.1
sqlalchemy>=2.0.36
psycopg2-binary>=2.9.10
pydantic>=2.10.1
pydantic-settings>=2.0.0
langchain>=0.3.7
openai>=1.55.0
anthropic>=0.39.0
PyYAML>=6.0.2
# Testing dependencies
pytest>=7.4.0
httpx>=0.24.1
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
requests>=2.31.0
beautifulsoup4>=4.12.2
aiohttp>=3.9.1
